Don't access the network in Python setup
This won't work if the build process is sandboxed.

The correct time to download the base README is when you prepare a release.
